refactor(app): move useRobot to redux-resources
Turns out we have yet another kind of resource: a redux resource! This
is a react hook that lightly wraps redux, which doesn't otherwise have
react integration. Isn't that cool?

Also, this useRobot hook is used outside Devices, and therefore must
live there.
