+++ app/src/organisms/Breadcrumbs/index.tsx
@@ -18,15 +18,13 @@ import {
 } from '@opentrons/components'
 import { ApiHostProvider } from '@opentrons/react-api-client'
 
-import {
-  useRobot,
-  useRunCreatedAtTimestamp,
-} from '/app/organisms/Devices/hooks'
+import { useRunCreatedAtTimestamp } from '/app/organisms/Devices/hooks'
 import { getProtocolDisplayName } from '/app/organisms/ProtocolsLanding/utils'
 import { getIsOnDevice } from '/app/redux/config'
 import { OPENTRONS_USB } from '/app/redux/discovery'
 import { getStoredProtocol } from '/app/redux/protocol-storage'
 import { appShellRequestor } from '/app/redux/shell/remote'
+import { useRobot } from '/app/redux-resources/robots'
 
 import type { DesktopRouteParams } from '../../App/types'
 import type { State } from '/app/redux/types'
