+++ app/src/organisms/Devices/ProtocolRun/ProtocolRunHeader/RunHeaderContent/ActionButton/index.tsx
@@ -17,11 +17,11 @@ import {
 import {
   useModuleCalibrationStatus,
   useProtocolDetailsForRun,
-  useRobot,
   useRobotAnalyticsData,
   useRunCalibrationStatus,
   useUnmatchedModulesForProtocol,
 } from '../../../../hooks'
+import { useRobot } from '/app/redux-resources/robots'
 import { useCurrentRunId } from '/app/resources/runs'
 import { useActionBtnDisabledUtils, useActionButtonProperties } from './hooks'
 import { getFallbackRobotSerialNumber, isRunAgainStatus } from '../../utils'
