+++ app/src/organisms/Devices/ProtocolRun/SetupModuleAndDeck/SetupModulesList.tsx
@@ -34,6 +34,7 @@ import {
   TC_MODULE_LOCATION_OT3,
 } from '@opentrons/shared-data'
 
+import { useRobot } from '/app/redux-resources/robots'
 import { TertiaryButton } from '/app/atoms/buttons'
 import { StatusLabel } from '/app/atoms/StatusLabel'
 import { useChainLiveCommands } from '/app/resources/runs'
@@ -44,7 +45,6 @@ import { getModuleTooHot } from '../../getModuleTooHot'
 import {
   useIsFlex,
   useModuleRenderInfoForProtocolById,
-  useRobot,
   useUnmatchedModulesForProtocol,
   useRunCalibrationStatus,
 } from '../../hooks'
