+++ app/src/organisms/Devices/RobotSettings/AdvancedTab/RobotInformation.tsx
@@ -9,7 +9,7 @@ import {
   LegacyStyledText,
   TYPOGRAPHY,
 } from '@opentrons/components'
-import { useRobot } from '/app/organisms/Devices/hooks'
+import { useRobot } from '/app/redux-resources/robots'
 import {
   getRobotSerialNumber,
   getRobotFirmwareVersion,
