+++ app/src/organisms/Devices/RobotSettings/AdvancedTab/RobotServerVersion.tsx
@@ -17,8 +17,9 @@ import { TertiaryButton } from '/app/atoms/buttons'
 import { getRobotApiVersion } from '/app/redux/discovery'
 import { getRobotUpdateDisplayInfo } from '/app/redux/robot-update'
 import { UpdateRobotBanner } from '../../../UpdateRobotBanner'
-import { useIsFlex, useRobot } from '../../hooks'
+import { useIsFlex } from '../../hooks'
 import { handleUpdateBuildroot } from '../UpdateBuildroot'
+import { useRobot } from '/app/redux-resources/robots'
 
 import type { State } from '/app/redux/types'
