+++ app/src/organisms/Devices/RobotSettings/AdvancedTab/__tests__/RobotInformation.test.tsx
@@ -10,11 +10,11 @@ import {
   getRobotFirmwareVersion,
   getRobotProtocolApiVersion,
 } from '/app/redux/discovery'
-import { useRobot } from '../../../hooks'
+import { useRobot } from '/app/redux-resources/robots'
 import { mockConnectableRobot } from '/app/redux/discovery/__fixtures__'
 import { RobotInformation } from '../RobotInformation'
 
-vi.mock('../../../hooks')
+vi.mock('/app/redux-resources/robots')
 vi.mock('/app/redux/discovery/selectors')
 
 const MOCK_ROBOT_SERIAL_NUMBER = '0.0.0'
