+++ app/src/organisms/Devices/RobotSettings/AdvancedTab/__tests__/RobotServerVersion.test.tsx
@@ -8,11 +8,11 @@ import { i18n } from '/app/i18n'
 import { getRobotApiVersion } from '/app/redux/discovery'
 import { getRobotUpdateDisplayInfo } from '/app/redux/robot-update'
 import { mockConnectableRobot } from '/app/redux/discovery/__fixtures__'
-import { useRobot } from '../../../hooks'
 import { handleUpdateBuildroot } from '../../UpdateBuildroot'
 import { RobotServerVersion } from '../RobotServerVersion'
+import { useRobot } from '/app/redux-resources/robots'
 
-vi.mock('../../../hooks')
+vi.mock('/app/redux-resources/robots')
 vi.mock('/app/redux/robot-update/selectors')
 vi.mock('/app/redux/discovery/selectors')
 vi.mock('../../UpdateBuildroot')
