+++ app/src/organisms/Devices/RobotSettings/ConnectNetwork/__tests__/DisconnectModal.test.tsx
@@ -6,7 +6,7 @@ import { renderWithProviders } from '/app/__testing-utils__'
 import { when } from 'vitest-when'
 
 import { i18n } from '/app/i18n'
-import { useRobot } from '/app/organisms/Devices/hooks'
+import { useRobot } from '/app/redux-resources/robots'
 import { useWifiList } from '/app/resources/networking/hooks'
 import {
   mockConnectableRobot,
@@ -34,7 +34,7 @@ import type { RequestState } from '/app/redux/robot-api/types'
 import type { State } from '/app/redux/types'
 
 vi.mock('/app/resources/networking/hooks')
-vi.mock('/app/organisms/Devices/hooks')
+vi.mock('/app/redux-resources/robots')
 vi.mock('/app/redux/networking')
 vi.mock('/app/redux/robot-api')
