+++ app/src/organisms/Devices/RobotSettings/RobotSettingsAdvanced.tsx
@@ -14,7 +14,7 @@ import {
 
 import { Divider } from '/app/atoms/structure'
 import { ToggleButton } from '/app/atoms/buttons'
-import { useIsFlex, useIsRobotBusy, useRobot } from '../hooks'
+import { useIsFlex, useIsRobotBusy } from '../hooks'
 import {
   DeviceReset,
   DisplayRobotName,
@@ -31,6 +31,7 @@ import {
   UsageSettings,
   UseOlderAspirateBehavior,
 } from './AdvancedTab'
+import { useRobot } from '/app/redux-resources/robots'
 import {
   updateSetting,
   getRobotSettings,
