+++ app/src/organisms/Devices/__tests__/HistoricalProtocolRunOverflowMenu.test.tsx
@@ -10,7 +10,8 @@ import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
 import { mockConnectableRobot } from '/app/redux/discovery/__fixtures__'
 import runRecord from '../ProtocolRun/ProtocolRunHeader/RunHeaderModalContainer/modals/__fixtures__/runRecord.json'
-import { useDownloadRunLog, useTrackProtocolRunEvent, useRobot } from '../hooks'
+import { useDownloadRunLog, useTrackProtocolRunEvent } from '../hooks'
+import { useRobot } from '/app/redux-resources/robots'
 import { useRunControls } from '../../RunTimeControl/hooks'
 import {
   useTrackEvent,
@@ -26,7 +27,8 @@ import type { CommandsData } from '@opentrons/api-client'
 
 vi.mock('/app/redux/analytics')
 vi.mock('/app/redux/robot-update/selectors')
-vi.mock('../../Devices/hooks')
+vi.mock('/app/redux-resources/robots')
+vi.mock('/app/organisms/Devices/hooks')
 vi.mock('../../RunTimeControl/hooks')
 vi.mock('/app/redux/analytics')
 vi.mock('/app/redux/config')
