+++ app/src/organisms/Devices/hooks/index.ts
@@ -17,7 +17,6 @@ export * from './useModuleCalibrationStatus'
 export * from './usePipetteOffsetCalibrations'
 export * from './usePipetteOffsetCalibration'
 export * from './useProtocolDetailsForRun'
-export * from './useRobot'
 export * from './useRobotType'
 export * from './useRunCalibrationStatus'
 export * from './useRunCreatedAtTimestamp'
