+++ app/src/organisms/Devices/hooks/usePipetteOffsetCalibration.ts
@@ -6,7 +6,7 @@ import {
   fetchPipetteOffsetCalibrations,
 } from '/app/redux/calibration'
 import { useDispatchApiRequest } from '/app/redux/robot-api'
-import { useRobot } from '.'
+import { useRobot } from '/app/redux-resources/robots'
 
 import type { PipetteOffsetCalibration } from '/app/redux/calibration/types'
 import type { State } from '/app/redux/types'
