+++ app/src/organisms/Devices/hooks/useTrackCreateProtocolRunEvent.ts
@@ -3,7 +3,7 @@ import { parseProtocolRunAnalyticsData } from './useProtocolRunAnalyticsData'
 import { parseProtocolAnalysisOutput } from './useStoredProtocolAnalysis'
 
 import type { StoredProtocolData } from '/app/redux/protocol-storage'
-import { useRobot } from './useRobot'
+import { useRobot } from '/app/redux-resources/robots'
 
 type CreateProtocolRunEventName =
   | 'createProtocolRecordRequest'
