+++ app/src/organisms/Devices/hooks/useUnmatchedModulesForProtocol.ts
@@ -2,11 +2,8 @@ import reduce from 'lodash/reduce'
 
 import type { AttachedModule } from '/app/redux/modules/types'
 
-import {
-  useAttachedModules,
-  useModuleRenderInfoForProtocolById,
-  useRobot,
-} from '.'
+import { useRobot } from '/app/redux-resources/robots'
+import { useAttachedModules, useModuleRenderInfoForProtocolById } from '.'
 import {
   NON_CONNECTING_MODULE_TYPES,
   getModuleType,
