+++ app/src/organisms/RobotSettingsCalibration/CalibrationDataDownload.tsx
@@ -22,13 +22,13 @@ import {
   useDeckCalibrationData,
   useIsFlex,
   usePipetteOffsetCalibrations,
-  useRobot,
   useTipLengthCalibrations,
 } from '/app/organisms/Devices/hooks'
 import {
   useTrackEvent,
   ANALYTICS_CALIBRATION_DATA_DOWNLOADED,
 } from '/app/redux/analytics'
+import { useRobot } from '/app/redux-resources/robots'
 import { useIsEstopNotDisengaged } from '/app/resources/devices/hooks/useIsEstopNotDisengaged'
 
 // TODO(bc, 2022-02-08): replace with support article when available
