+++ app/src/organisms/RobotSettingsCalibration/RobotSettingsDeckCalibration.tsx
@@ -12,7 +12,8 @@ import {
 } from '@opentrons/components'
 
 import { formatLastModified } from '/app/organisms/CalibrationPanels/utils'
-import { useDeckCalibrationData, useRobot } from '/app/organisms/Devices/hooks'
+import { useDeckCalibrationData } from '/app/organisms/Devices/hooks'
+import { useRobot } from '/app/redux-resources/robots'
 import * as RobotApi from '/app/redux/robot-api'
 
 import type { State } from '/app/redux/types'
