+++ app/src/organisms/RobotSettingsCalibration/__tests__/CalibrationDataDownload.test.tsx
@@ -39,9 +39,9 @@ import {
   useDeckCalibrationData,
   useIsFlex,
   usePipetteOffsetCalibrations,
-  useRobot,
   useTipLengthCalibrations,
 } from '/app/organisms/Devices/hooks'
+import { useRobot } from '/app/redux-resources/robots'
 import { renderWithProviders } from '/app/__testing-utils__'
 import { useIsEstopNotDisengaged } from '/app/resources/devices/hooks/useIsEstopNotDisengaged'
 import { CalibrationDataDownload } from '../CalibrationDataDownload'
@@ -57,6 +57,7 @@ vi.mock('file-saver', async importOriginal => {
 vi.mock('@opentrons/react-api-client')
 vi.mock('/app/redux/analytics')
 vi.mock('/app/organisms/Devices/hooks')
+vi.mock('/app/redux-resources/robots')
 vi.mock('/app/resources/devices/hooks/useIsEstopNotDisengaged')
 
 let mockTrackEvent: any
