+++ app/src/organisms/RobotSettingsCalibration/__tests__/RobotSettingsCalibration.test.tsx
@@ -22,11 +22,11 @@ import {
 import {
   useIsFlex,
   usePipetteOffsetCalibrations,
-  useRobot,
   useAttachedPipettes,
   useRunStatuses,
   useAttachedPipettesFromInstrumentsQuery,
 } from '/app/organisms/Devices/hooks'
+import { useRobot } from '/app/redux-resources/robots'
 
 import { CalibrationDataDownload } from '../CalibrationDataDownload'
 import { CalibrationHealthCheck } from '../CalibrationHealthCheck'
@@ -50,6 +50,7 @@ vi.mock('/app/organisms/CalibrationStatusCard')
 vi.mock('/app/redux/config')
 vi.mock('/app/redux/sessions/selectors')
 vi.mock('/app/redux/robot-api/selectors')
+vi.mock('/app/redux-resources/robots')
 vi.mock('/app/organisms/Devices/hooks')
 vi.mock('../CalibrationDataDownload')
 vi.mock('../CalibrationHealthCheck')
