+++ app/src/organisms/RobotSettingsCalibration/index.tsx
@@ -22,11 +22,11 @@ import { CalibrateDeck } from '/app/organisms/CalibrateDeck'
 import { CalibrationStatusCard } from '/app/organisms/CalibrationStatusCard'
 import { CheckCalibration } from '/app/organisms/CheckCalibration'
 import {
-  useRobot,
   useRunStatuses,
   useIsFlex,
   useAttachedPipettesFromInstrumentsQuery,
 } from '/app/organisms/Devices/hooks'
+import { useRobot } from '/app/redux-resources/robots'
 import { HowCalibrationWorksModal } from '/app/organisms/HowCalibrationWorksModal'
 import { CONNECTABLE } from '/app/redux/discovery'
 import * as RobotApi from '/app/redux/robot-api'
