+++ app/src/pages/Desktop/Devices/DeviceDetails/__tests__/DeviceDetails.test.tsx
@@ -7,12 +7,13 @@ import { MemoryRouter, Route, Routes } from 'react-router-dom'
 import { renderWithProviders } from '/app/__testing-utils__'
 
 import { i18n } from '/app/i18n'
-import { useRobot, useSyncRobotClock } from '/app/organisms/Devices/hooks'
+import { useSyncRobotClock } from '/app/organisms/Devices/hooks'
 import { InstrumentsAndModules } from '/app/organisms/Devices/InstrumentsAndModules'
 import { RecentProtocolRuns } from '/app/organisms/Devices/RecentProtocolRuns'
 import { RobotOverview } from '/app/organisms/Devices/RobotOverview'
 import { getScanning } from '/app/redux/discovery'
 import { mockConnectableRobot } from '/app/redux/discovery/__fixtures__'
+import { useRobot } from '/app/redux-resources/robots'
 import { DeviceDetails } from '..'
 
 import type { State } from '/app/redux/types'
@@ -22,6 +23,7 @@ vi.mock('/app/organisms/Devices/InstrumentsAndModules')
 vi.mock('/app/organisms/Devices/RecentProtocolRuns')
 vi.mock('/app/organisms/Devices/RobotOverview')
 vi.mock('/app/redux/discovery')
+vi.mock('/app/redux-resources/robots')
 
 const render = (path = '/') => {
   return renderWithProviders(
