+++ app/src/pages/Desktop/Devices/RobotSettings/index.tsx
@@ -26,7 +26,7 @@ import { appShellRequestor } from '/app/redux/shell/remote'
 import { getRobotUpdateSession } from '/app/redux/robot-update'
 import { getDevtoolsEnabled } from '/app/redux/config'
 import { Banner } from '/app/atoms/Banner'
-import { useRobot } from '/app/organisms/Devices/hooks'
+import { useRobot } from '/app/redux-resources/robots'
 import { Line } from '/app/atoms/structure'
 import { NavTab } from '/app/molecules/NavTab'
 import { RobotSettingsCalibration } from '/app/organisms/RobotSettingsCalibration'
