+++ app/src/redux-resources/robots/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useRobot'
