+++ app/src/redux-resources/robots/index.ts
@@ -1,0 +1,1 @@
+export * from './hooks'
