+++ app/src/resources/networking/hooks/useWifiList.ts
@@ -1,7 +1,7 @@
 import uniqBy from 'lodash/uniqBy'
 import orderBy from 'lodash/orderBy'
 import { useWifiQuery } from '@opentrons/react-api-client'
-import { useRobot } from '/app/organisms/Devices/hooks'
+import { useRobot } from '/app/redux-resources/robots'
 
 import type { WifiNetwork } from '@opentrons/api-client'
